# src/main/java/com/insight7/ordertracking/controller/models/Message.java
package com.insight7.ordertracking.controller.models;

import com.fasterxml.jackson.annotation.JsonInclude;
import lombok.Builder;
import lombok.Data;

@Data
@Builder
@JsonInclude(JsonInclude.Include.NON_NULL)
public class Message {
    private String statusCode;
    private String message;
}