# swagger.yaml
swagger: "2.0"
info:
  description: "This is a sample server Petstore server.  You can find out more about     Swagger at [http://swagger.io](http://swagger.io) or on [irc.freenode.net, #swagger](http://swagger.io/irc/).      For this sample, you can use the api key `special-key` to test the authorization     filters."
  version: "1.0.0"
  title: "Swagger Petstore"
  termsOfService: "http://swagger.io/terms/"
  contact:
    email: "apiteam@swagger.io"
  license:
    name: "Apache 2.0"
    url: "http://www.apache.org/licenses/LICENSE-2.0.html"
host: "petstore.swagger.io"
basePath: "/v2"
schemes:
  - "https"
  - "http"
paths:
  /orders:
    get:
      summary: "Get list of orders"
      description: ""
      produces:
        - "application/json"
      parameters:
        - name: "source"
          in: "query"
          description: "Filter order by source"
          type: "string"
        - name: "destination"
          in: "query"
          description: "Filter order by destination"
          type: "string"
        - name: "status"
          in: "query"
          description: "Filter order by status"
          type: "string"
          enum:
            - "PLACED"
            - "APPROVED"
            - "IN_TRANSIT"
            - "DELIVERED"
        - name: "limit"
          in: "query"
          description: "Specifies the maximun number of results to return"
          type: "integer"
          format: "int32"
        - name: "offset"
          in: "query"
          description: "Offset of the results value"
          type: "integer"
          format: "int32"
      responses:
        200:
          description: "Successful Response"
          schema:
            $ref: "#/definitions/Orders"
    post:
      summary: "Create new Order"
      description: ""
      consumes:
        - "application/json"
      produces:
        - "application/json"
      parameters:
        - name: "body"
          in: "body"
          description: "Create a new order"
          required: true
          schema:
            $ref: "#/definitions/CreateOrderRequest"
      responses:
        200:
          description: "Order created successfully"
          schema:
            $ref: "#/definitions/Order"
  /orders/{orderId}:
    get:
      summary: "Get order for the given order id"
      description: ""
      produces:
        - "application/json"
      parameters:
        - name: "orderId"
          in: "path"
          description: "Order id for the required order "
          required: true
          type: "integer"
          format: "int32"
      responses:
        200:
          description: "successful Response"
          schema:
            $ref: "#/definitions/Order"
    put:
      summary: "Update an order"
      description: ""
      produces:
        - "application/json"
      consumes:
        - "application/json"
      parameters:
        - name: "orderId"
          in: "path"
          description: "Order id for the required order "
          required: true
          type: "string"
        - name: "body"
          in: "body"
          description: "New order details"
          schema:
            $ref: "#/definitions/UpdateOrderRequest"
          required: true
      responses:
        200:
          description: "successful Response"
          schema:
            $ref: "#/definitions/Order"
    delete:
      summary: "Delete the order"
      description: ""
      produces:
        - "application/json"
      parameters:
        - name: "orderId"
          in: "path"
          description: "Order id for the required order "
          required: true
          type: "string"
      responses:
        200:
          description: "successful Response"
          schema:
            $ref: "#/definitions/Message"
  /tracking/{orderId}:
    get:
      summary: "Get tracking for the given order id"
      description: ""
      produces:
        - "application/json"
      parameters:
        - name: "orderId"
          in: "path"
          description: "Order id for the required order tracking."
          required: true
          type: "string"
      responses:
        200:
          description: "successful Response"
          schema:
            $ref: "#/definitions/Trackings"
    post:
      summary: "Create a new tracking entry for that order"
      description: ""
      produces:
        - "application/json"
      consumes:
        - "application/json"
      parameters:
        - name: "orderId"
          in: "path"
          description: "Order id for the required order tracking."
          required: true
          type: "string"
        - name: "body"
          in: "body"
          description: "New tracking details"
          schema:
            $ref: "#/definitions/CreateTrackingRequest"
          required: true
      responses:
        200:
          description: "successful Response"
          schema:
            $ref: "#/definitions/Tracking"
  /tracking/{orderId}/{trackingId}:
    get:
      summary: "Get tracking details for the given order id and tracking id"
      description: ""
      produces:
        - "application/json"
      consumes:
        - "application/json"
      parameters:
        - name: "orderId"
          in: "path"
          description: "Order id for the required order tracking."
          required: true
          type: "string"
        - name: "trackingId"
          in: "path"
          description: "tracking id for the required order tracking."
          required: true
          type: "string"
      responses:
        200:
          description: "successful Response"
          schema:
            $ref: "#/definitions/Tracking"
    put:
      summary: "Update tracking details for the given order id and tracking id"
      description: ""
      produces:
        - "application/json"
      parameters:
        - name: "orderId"
          in: "path"
          description: "Order id for the required order tracking."
          required: true
          type: "string"
        - name: "trackingId"
          in: "path"
          description: "tracking id for the required order tracking."
          required: true
          type: "string"
        - name: "body"
          in: "body"
          description: "New tracking details"
          schema:
            $ref: "#/definitions/CreateTrackingRequest"
          required: true
      responses:
        200:
          description: "successful Response"
          schema:
            $ref: "#/definitions/Tracking"
definitions:
  Trackings:
    properties:
      pagination:
        type: "object"
        properties:
          count:
            type: "integer"
            format: "int32"
      order:
        $ref: "#/definitions/Order"
      items:
        type: "array"
        items:
          $ref: "#/definitions/Tracking"
  Orders:
    type: "object"
    properties:
      pagination:
        type: "object"
        properties:
          count:
            type: "integer"
            format: "int32"
          limit:
            type: "integer"
            format: "int32"
          offset:
            type: "integer"
            format: "int32"
          total:
            type: "integer"
            format: "int32"
      items:
        type: "array"
        items:
          $ref: "#/definitions/Order"
  Order:
    type: "object"
    properties:
      id:
        type: "integer"
        format: "int32"
      source:
        type: "string"
      destination:
        type: "string"
      status:
        type: "string"
        description: "Order Status"
        enum:
          - "PLACED"
          - "APPROVED"
          - "IN_TRANSIT"
          - "DELIVERED"
      createTime:
        type: "string"
        format: "date-time"
      updateTime:
        type: "string"
        format: "date-time"
      deliveryTime:
        type: "string"
        format: "date-time"
      noOfPkgs:
        type: "integer"
        format: "int32"
      weight:
        type: "integer"
        format: "int32"
      volume:
        type: "integer"
        format: "int32"
  Tracking:
    properties:
      id:
        type: "integer"
        format: "int32"
      updateTime:
        type: "string"
        format: "date-time"
      location:
        type: "string"
      message:
        type: "string"
  CreateOrderRequest:
    properties:
      source:
        type: "string"
      destination:
        type: "string"
      noOfPkgs:
        type: "integer"
        format: "int32"
      weight:
        type: "integer"
        format: "int32"
      volume:
        type: "integer"
        format: "int32"
  UpdateOrderRequest:
    properties:
      destination:
        type: "string"
      status:
        type: "string"
        description: "Order Status"
        enum:
          - "PLACED"
          - "APPROVED"
          - "IN_TRANSIT"
          - "DELIVERED"
      noOfPkgs:
        type: "integer"
        format: "int32"
      weight:
        type: "integer"
        format: "int32"
      volume:
        type: "integer"
        format: "int32"
  CreateTrackingRequest:
    properties:
      location:
        type: "string"
      message:
        type: "string"
  Message:
    properties:
      statusCode:
        type: "string"
      message:
        type: "string"